Rewrite I/O code
The domain model supported by the original I/O code differed too much
from what I would like to use myself.

Furthermore the original code was a bit wasteful with resources, was
too lenient in its use of unsafe, and was too aggressive in its error
handling (panicing instead of giving me a chance to log something to a
file).